Make sub_books generic
Instead of hard-coding the books the script now walks through
all existing subdirectories and looks for mkdocs.yml files.
All subdirs except the one marked as the main book are now
considered as sub_books, making it easy to use the script for
new sites or to add new sub_books.